# M4-Summative-Aynalem-Sandy-Hanan/openapi (4).yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /game/{id}:
    get:
      tags:
        - game-controller
      operationId: getGameById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GameViewModel'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
    put:
      tags:
        - game-controller
      operationId: updateGame
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameViewModel'
        required: true
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
    delete:
      tags:
        - game-controller
      operationId: deleteGame
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /console/{id}:
    get:
      tags:
        - console-controller
      operationId: getConsolById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ConsoleViewModel'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
    put:
      tags:
        - console-controller
      operationId: updateConsole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsoleViewModel'
        required: true
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
    delete:
      tags:
        - console-controller
      operationId: deleteConsole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
...
  /game:
    get:
      tags:
        - game-controller
      operationId: getAllGames
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameViewModel'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
    post:
      tags:
        - game-controller
      operationId: addGame
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameViewModel'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GameViewModel'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /console:
    get:
      tags:
        - console-controller
      operationId: getAllConsoles
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsoleViewModel'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
    post:
      tags:
        - console-controller
      operationId: postConsole
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsoleViewModel'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ConsoleViewModel'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
...
  /manufacturer/{manufacturer}:
    get:
      tags:
        - console-controller
      operationId: getConsoleByManufacturer
      parameters:
        - name: manufacturer
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsoleViewModel'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /game/title/{title}:
    get:
      tags:
        - game-controller
      operationId: findGamesByTitle
      parameters:
        - name: title
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameViewModel'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /game/studio/{studio}:
    get:
      tags:
        - game-controller
      operationId: findGamesByStudio
      parameters:
        - name: studio
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameViewModel'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /game/esrb/{esrbRating}:
    get:
      tags:
        - game-controller
      operationId: findGamesByRating
      parameters:
        - name: esrbRating
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameViewModel'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
components:
  schemas:
    CustomErrorResponse:
      type: object
      properties:
        errorMsg:
          type: string
        status:
          type: integer
          format: int32
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time
...
    GameViewModel:
      type: object
      properties:
        gameId:
          type: integer
          format: int32
        studio:
          type: string
        esrbRating:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer
          format: int32
    ConsoleViewModel:
      type: object
      properties:
        consoleId:
          type: integer
          format: int32
        consoleManufacturer:
          type: string
        consoleModel:
          type: string
        consolePrice:
          type: number
          format: double
        consoleQuantity:
          type: integer
          format: int32
        consoleProcessor:
          type: string
        consoleMemory:
          type: string